Use PS0 to start observing a command instead of the strcpy-hack.
Previously we checked bash's current_command_number (loaded via dlsym)
after every strcpy library-call in order to know when bash started
a new command execution. This has proven unreliable for bash 5.1.8 on
openSUSE Tumbleweed. When a subshell (echo forked; echo hi > foo)
was executed, the first strcpy did not fire yet.
Therefore we switched to PS0 calling «fork» where we start the
observation. Doing so we got rid of the dlsym-call altogether.
Further we make extra-sure, to not start tracking from a subshell,
which occurred in the above mentioned bash-OS-combo.